requirements.txt
asgiref==3.8.1
dataclasses==0.6
Django==4.2.16
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
drf-yasg==1.21.8
inflection==0.5.1
mongoengine==0.29.1
mysqlclient==2.2.7
packaging==24.2
PyJWT==2.10.1
pymongo==4.10.1
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
